Updated ArmController to use _t types Fixed bug in buffer size

+++ src/MotherboardFirmware/src/ArmController.h
@@ -6,28 +6,28 @@
 
 class ArmController {
 public:
-    static const int COUNTS_PER_REVOLUTION = 4;
-    static const int BUFFER_MULT = 4;
-    static const int BUFFER_SIZE = BUFFER_MULT * COUNTS_PER_REVOLUTION;
+    static const uint8_t COUNTS_PER_REVOLUTION = 4;
+    static const uint8_t BUFFER_MULT = 4;
+    static const uint8_t BUFFER_SIZE = BUFFER_MULT * COUNTS_PER_REVOLUTION;
 
-    ArmController(int pwmPin, int hallPin);
+    ArmController(uint8_t pwmPin, uint8_t hallPin);
 
     void setup();
     void setThrottle(float throttle); // 0.0 to 1.0
     void stop();
 
-    float getRPM(int samples = 10);
+    float getRPM(uint8_t samples = 10);
     void resetHallTimes();
     
 private:
-    int _pwmPin;
-    int _hallPin;
+    uint8_t _pwmPin;
+    uint8_t _hallPin;
     
     // Volatiles for Interrupt
-    volatile int _validHallSamples;
-    volatile unsigned long _hallTimes[BUFFER_SIZE];
-    volatile int _hallIndex;
-    volatile unsigned long _lastPulseTime;
+    volatile uint8_t _validHallSamples;
+    volatile uint32_t _hallTimes[BUFFER_SIZE];
+    volatile uint8_t _hallIndex;
+    volatile uint32_t _lastPulseTime;
 
     float _throttle;
     Servo _escServo;
